Cleaning up code action menu code (#160724)
- Move keybinding resolver to own file
- Use list methods instead of re-implementing them
- Reduce duplicated use state
